move request handlers for boards to separate file
